requirements.txt
urllib3==1.26.15
docker==3.6.0
PyYAML==6.0
redis==4.4.2
python-logging-loki==0.3.1
boto3==1.26.54
requests-mock==1.10.0
mock==5.0.1
kubernetes==25.3.0
google-cloud-compute==1.9.0
